+++ lib/receipt_verifier/client.ex
@@ -2,7 +2,6 @@ defmodule ReceiptVerifier.Client do
   @moduledoc false
 
   alias ReceiptVerifier.Error
-  alias Poison.Parser, as: JSONParser
 
   @endpoints [
     production: 'https://buy.itunes.apple.com/verifyReceipt',
@@ -12,31 +11,35 @@ defmodule ReceiptVerifier.Client do
   @doc """
   Send the iTunes receipt to Apple Store, and parse the response as map
   """
-  @spec request(String.t, ReceiptVerifier.options) :: {:ok, map} | {:error, any}
-  def request(receipt, opts \\ []) do
-    with(
-      {:ok, {{_, 200, _}, _, body}} <- do_request(receipt, opts),
-      {:ok, json} <- JSONParser.parse(body)
-    ) do
+  @spec request(String.t(), map) :: {:ok, map} | {:error, any}
+  def request(receipt, opts) do
+    with {:ok, {{_, 200, _}, _, body}} <- do_request(receipt, opts),
+         {:ok, json} <- Poison.decode(body) do
       {:ok, json}
     else
+      {:ok, {{_, status_code, msg}, _, _body}} ->
+        {:error, %Error{code: status_code, message: msg}}
+
       {:error, :invalid} ->
         # Poison error
         {:error, %Error{code: 502, message: "The response from Apple's Server is malformed"}}
+
       {:error, {:invalid, msg}} ->
         # Poison error
-        {:error, %Error{code: 502, message: "The response from Apple's Server is malformed: #{msg}"}}
+        {:error,
+         %Error{code: 502, message: "The response from Apple's Server is malformed: #{msg}"}}
+
       {:error, reason} ->
         {:error, reason}
     end
   end
 
   defp do_request(receipt, opts) do
-    env = Keyword.get(opts, :env, :production)
-    url = get_endpoint_url(env)
+    url = get_endpoint_url(opts.env)
 
     request_body = prepare_request_body(receipt, opts)
     content_type = 'application/json'
+
     request_headers = [
       {'Accept', 'application/json'}
     ]
@@ -49,19 +52,25 @@ defmodule ReceiptVerifier.Client do
     |> Keyword.get(env)
   end
 
+  defp get_endpoint_url(:auto) do
+    @endpoints
+    |> Keyword.get(:production)
+  end
+
   defp prepare_request_body(receipt, opts) do
     %{
       "receipt-data" => receipt
     }
     |> maybe_set_password(opts)
     |> maybe_set_exclude_old_transactions(opts)
-    |> Poison.encode!
+    |> Poison.encode!()
   end
 
   defp maybe_set_password(data, opts) do
-    case Keyword.get(opts, :password) do
+    case Map.get(opts, :password) do
       nil ->
         data
+
       password ->
         data
         |> Map.put("password", password)
@@ -69,9 +78,10 @@ defmodule ReceiptVerifier.Client do
   end
 
   defp maybe_set_exclude_old_transactions(data, opts) do
-    case Keyword.get(opts, :exclude_old_transactions) do
+    case Map.get(opts, :exclude_old_transactions) do
       nil ->
         data
+
       exclude_old_transactions ->
         data
         |> Map.put("exclude-old-transactions", exclude_old_transactions)